fix: some translations which has different locale may has one textId
